fix queries.

+++ src/local/butler/scripts/grouper_queries.py
@@ -28,43 +28,57 @@ def execute(args):
     os.mkdir(storage_dir)
 
   # Check the amount of testcases fixed/NA/not fixed in groups
-  groups_with_fixed_testcases = set()
+  groups_with_fixed_testcases = set()  
+  fixed_file = os.path.join(storage_dir, 'groups_with_fixed.pkl')
   groups_with_na_testcases = set()
+  na_file = os.path.join(storage_dir, 'groups_with_na.pkl')
   groups_with_not_fixed_testcases = set()
-  grouped_testcases = data_types.Testcase.query(data_types.Testcase.group_id != 0)
-  count = 0
-  for testcase in grouped_testcases:
-    if not testcase.fixed:
-      groups_with_not_fixed_testcases.add(testcase.group_id)
-    elif testcase.fixed == 'NA':
-      groups_with_na_testcases.add(testcase.group_id)
-    else:
-      groups_with_fixed_testcases.add(testcase.group_id)
-    count += 1
-    if count % 100 == 0:
-      print(f'{count} testcases analyzed.')
+  not_fixed_file = os.path.join(storage_dir, 'groups_with_not_fixed.pkl')
 
-  with open(os.path.join(storage_dir, 'groups_with_fixed.pkl'), 'wb') as f:
-    pickle.dump(groups_with_fixed_testcases, f)
+  if not os.path.exists(fixed_file):
+    grouped_testcases = data_types.Testcase.query(data_types.Testcase.group_id != 0)
+    count = 0
+    for testcase in grouped_testcases:
+      if not testcase.fixed:
+        groups_with_not_fixed_testcases.add(testcase.group_id)
+      elif testcase.fixed == 'NA':
+        groups_with_na_testcases.add(testcase.group_id)
+      else:
+        groups_with_fixed_testcases.add(testcase.group_id)
+      count += 1
+      if count % 100 == 0:
+        print(f'{count} testcases analyzed.')
 
-  with open(os.path.join(storage_dir, 'groups_with_na.pkl'), 'wb') as f:
-    pickle.dump(groups_with_na_testcases, f)
+    with open(fixed_file, 'wb') as f:
+      pickle.dump(groups_with_fixed_testcases, f)
 
-  with open(os.path.join(storage_dir, 'groups_with_not_fixed.pkl'), 'wb') as f:
-    pickle.dump(groups_with_not_fixed_testcases, f)
+    with open(na_file, 'wb') as f:
+      pickle.dump(groups_with_na_testcases, f)
 
+    with open(not_fixed_file, 'wb') as f:
+      pickle.dump(groups_with_not_fixed_testcases, f)
+  else:
+    print(f'Loading from existing files.')
+    with open(fixed_file, 'rb') as f:
+      groups_with_fixed_testcases = pickle.load(f)
+
+    with open(na_file, 'rb') as f:
+      groups_with_na_testcases = pickle.load(f)
+
+    with open(not_fixed_file, 'rb') as f:
+      groups_with_not_fixed_testcases = pickle.load(f)
 
   only_fixed = groups_with_fixed_testcases.difference(groups_with_na_testcases).difference(groups_with_not_fixed_testcases)
   only_na = groups_with_na_testcases.difference(groups_with_fixed_testcases).difference(groups_with_not_fixed_testcases)
   only_not_fixed = groups_with_not_fixed_testcases.difference(groups_with_fixed_testcases).difference(groups_with_na_testcases)
-  print(f'# Groups with only fixed: {only_fixed}')
-  print(f'# Groups with only NA: {only_na}')
-  print(f'# Groups with only not fixed: {only_not_fixed}')
+  print(f'# Groups with only fixed: {len(only_fixed)}')
+  print(f'# Groups with only NA: {len(only_na)}')
+  print(f'# Groups with only not fixed: {len(only_not_fixed)}')
 
 
   fixed_and_not_fixed = groups_with_fixed_testcases.intersection(groups_with_not_fixed_testcases)
   fixed_and_na = groups_with_fixed_testcases.intersection(groups_with_na_testcases)
   not_fixed_and_na = groups_with_not_fixed_testcases.intersection(groups_with_na_testcases)
-  print(f'# Groups with fixed and not fixed: {fixed_and_not_fixed}')
-  print(f'# Groups with fixed and NA: {fixed_and_na}')
-  print(f'# Groups with not fixed and NA: {not_fixed_and_na}')
+  print(f'# Groups with fixed and not fixed: {len(fixed_and_not_fixed)}')
+  print(f'# Groups with fixed and NA: {len(fixed_and_na)}')
+  print(f'# Groups with not fixed and NA: {len(not_fixed_and_na)}')
